+++ packages/retry/src/tests/retry.test.js
@@ -2,122 +2,137 @@ import { UPLOADER_EVENTS } from "@rpldy/uploader";
 import { RETRY_EVENT } from "../consts";
 import retryEnhancer from "../retry";
 
-describe("", () => {
+describe("retry tests", () => {
 
-    const items = [
-        { id: "f-1", file: "file1", batchId: "b1" },
-        { id: "f-2", url: "url2", batchId: "b1" },
-        { id: "f-3", file: "file3", batchId: "b2" },
-        { id: "f-4", file: "file4", batchId: "b2" },
-        { id: "f-5", url: "url5", batchId: "b3" },
-    ];
+	const items = [
+		{ id: "f-1", file: "file1", batchId: "b1" },
+		{ id: "f-2", url: "url2", batchId: "b1" },
+		{ id: "f-3", file: "file3", batchId: "b2" },
+		{ id: "f-4", file: "file4", batchId: "b2" },
+		{ id: "f-5", url: "url5", batchId: "b3" },
+	];
 
-    const getTestRetry = () => {
+	const getTestRetry = (itemEvent) => {
+		const trigger = jest.fn();
+		let retry, retryBatch;
 
-        const trigger = jest.fn();
-        let retry, retryBatch;
+		let uploader = {
+			registerExtension: (name, methods) => {
+				retry = methods.retry;
+				retryBatch = methods.retryBatch;
+			},
 
-        let uploader = {
-            registerExtension: (name, methods) => {
-                retry = methods.retry;
-                retryBatch = methods.retryBatch;
-            },
+			on: (event, method) => {
+				if (event === itemEvent) {
+					items.forEach(method);
+				}
+			},
 
-            on: (event, method) => {
-                if (event === UPLOADER_EVENTS.ITEM_ERROR) {
-                    items.forEach(method);
-                }
-            },
+			add: jest.fn(),
+		};
 
-            add: jest.fn(),
-        };
+		uploader = retryEnhancer(uploader, trigger);
 
-        uploader = retryEnhancer(uploader, trigger);
+		return {
+			retry,
+			retryBatch,
+			uploader,
+			trigger,
+		};
+	};
 
-        return {
-            retry,
-            retryBatch,
-            uploader,
-            trigger,
-        };
-    };
+	describe("retry all tests", () => {
+		it.each([
+			UPLOADER_EVENTS.ITEM_ERROR,
+			UPLOADER_EVENTS.ITEM_ABORT,
+		])("should send all items to retry ", (itemEvent) => {
 
+			const { retry, trigger, uploader, } = getTestRetry(itemEvent);
 
-    describe("retry all tests", () => {
+			const result = retry();
 
-        it("should send all items to retry ", () => {
+			expect(result).toBe(true);
 
-            const { retry, trigger, uploader, } = getTestRetry();
+			const options = {autoUpload: true};
 
-            const result = retry();
+			expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items , options});
+			expect(uploader.add).toHaveBeenCalledWith(items, options);
 
-            expect(result).toBe(true);
+			const result2 = retry();
+			expect(result2).toBe(false);
 
-            const expectedItems = items.map((i) => i.file || i.url);
+			expect(trigger).toHaveBeenCalledTimes(1);
+			expect(uploader.add).toHaveBeenCalledTimes(1);
+		});
 
-            expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items: expectedItems });
-            expect(uploader.add).toHaveBeenCalledWith(expectedItems, undefined);
+		it.each([
+			UPLOADER_EVENTS.ITEM_ERROR,
+			UPLOADER_EVENTS.ITEM_ABORT,
+		])("should send all items with options", (itemEvent) => {
+			const { retry, trigger, uploader, } = getTestRetry(itemEvent);
+			const options = { autoUpload: false };
 
-            const result2 = retry();
-            expect(result2).toBe(false);
+			retry(null, options);
 
-            expect(trigger).toHaveBeenCalledTimes(1);
-            expect(uploader.add).toHaveBeenCalledTimes(1);
-        });
+			expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items, options });
+			expect(uploader.add).toHaveBeenCalledWith(items, options);
+		});
+	});
 
-        it("should send all items with options", () => {
-            const { retry, trigger, uploader, } = getTestRetry();
-            const options = { foo: "bar" };
+	describe("retry item tests", () => {
+		it.each([
+			UPLOADER_EVENTS.ITEM_ERROR,
+			UPLOADER_EVENTS.ITEM_ABORT,
+		])("should send requested item to retry", (itemEvent) => {
+			const { retry, trigger, uploader, } = getTestRetry(itemEvent);
+			const options = { foo: "bar" };
 
-            retry(null, options);
+			const result = retry("f-3", options);
+			expect(result).toBe(true);
 
-            const expectedItems = items.map((i) => i.file || i.url);
+			const expectedOptions = {
+				...options,
+				autoUpload: true,
+			};
 
-            expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items: expectedItems, options });
-            expect(uploader.add).toHaveBeenCalledWith(expectedItems, options);
-        });
-    });
+			expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items: [items[2]], options: expectedOptions });
+			expect(uploader.add).toHaveBeenCalledWith([items[2]], expectedOptions);
 
-    describe("retry item tests", () => {
+			const result2 = retry("f-3", options);
+			expect(result2).toBe(false);
 
-        it("should send requested item to retry", () => {
-            const { retry, trigger, uploader, } = getTestRetry();
-            const options = { foo: "bar" };
+			expect(trigger).toHaveBeenCalledTimes(1);
+			expect(uploader.add).toHaveBeenCalledTimes(1);
+		});
+	});
 
-            const result = retry("f-3", options);
-            expect(result).toBe(true);
+	describe("retry batch tests", () => {
+		it.each([
+			UPLOADER_EVENTS.ITEM_ERROR,
+			UPLOADER_EVENTS.ITEM_ABORT,
+		])("should send batch items to retry ", (itemEvent) => {
+			const { retryBatch, trigger, uploader, } = getTestRetry(itemEvent);
 
-            expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items: [items[2].file], options });
-            expect(uploader.add).toHaveBeenCalledWith([items[2].file], options);
+			const options = { foo: "bar" };
 
-            const result2 = retry("f-3", options);
-            expect(result2).toBe(false);
+			const result = retryBatch("b2", options);
+			expect(result).toBe(true);
 
-            expect(trigger).toHaveBeenCalledTimes(1);
-            expect(uploader.add).toHaveBeenCalledTimes(1);
-        });
-    });
+			const expectedOptions = {
+				...options,
+				autoUpload: true,
+			};
 
-    describe("retry batch tests", () => {
+			const expectedItems =  [items[2], items[3]];
 
-        it("should send batch items to retry ", () => {
-            const { retryBatch, trigger, uploader, } = getTestRetry();
+			expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items: expectedItems, options: expectedOptions });
+			expect(uploader.add).toHaveBeenCalledWith(expectedItems, expectedOptions);
 
-            const options = { foo: "bar" };
+			const result2 = retryBatch("b2", options);
+			expect(result2).toBe(false);
 
-            const result = retryBatch("b2", options);
-            expect(result).toBe(true);
-
-            const expectedItems = [items[2], items[3]].map((i) => i.file || i.url);
-
-            expect(trigger).toHaveBeenCalledWith(RETRY_EVENT, { items: expectedItems, options });
-            expect(uploader.add).toHaveBeenCalledWith(expectedItems, options);
-
-            const result2 = retryBatch("b2", options);
-            expect(result2).toBe(false);
-
-            expect(trigger).toHaveBeenCalledTimes(1);
-            expect(uploader.add).toHaveBeenCalledTimes(1);
-        });
-    });
+			expect(trigger).toHaveBeenCalledTimes(1);
+			expect(uploader.add).toHaveBeenCalledTimes(1);
+		});
+	});
 });