+++ packages/shared/src/tests/batchItem.test.js
@@ -47,4 +47,42 @@ describe("create batchItem tests", () => {
         }).toThrow();
     });
 
+	it("should recycle file batch item", () => {
+
+		const file = { name: "test", type: "image/jpg" };
+		const fileItem = createItem(file, "b1");
+
+		fileItem.state = "DONE";
+		fileItem.completed = 100;
+		fileItem.loaded = 1000;
+		fileItem.aborted = true;
+
+		const recycled = createItem(fileItem, "b2");
+
+		expect(recycled.id).toBe(fileItem.id);
+		expect(recycled.state).toBe(FILE_STATES.ADDED);
+		expect(recycled.completed).toBe(0);
+		expect(recycled.loaded).toBe(0);
+		expect(recycled.aborted).toBe(false);
+		expect(recycled.file).toEqual(file);
+	});
+
+	it("should recycle url batch item", () => {
+
+		const fileItem = createItem( "file.com", "b1");
+
+		fileItem.state = "DONE";
+		fileItem.completed = 100;
+		fileItem.loaded = 1000;
+		fileItem.aborted = true;
+
+		const recycled = createItem(fileItem, "b2");
+
+		expect(recycled.id).toBe(fileItem.id);
+		expect(recycled.state).toBe(FILE_STATES.ADDED);
+		expect(recycled.completed).toBe(0);
+		expect(recycled.loaded).toBe(0);
+		expect(recycled.aborted).toBe(false);
+		expect(recycled.url).toBe("file.com");
+	});
 });