+++ packages/ui/upload-preview/src/types.js
@@ -6,7 +6,9 @@ import { PREVIEW_TYPES } from "./consts";
 export type PreviewType = $Values<typeof PREVIEW_TYPES>;
 
 export type PreviewData = {
+	id: string,
 	url: string,
+	name: string,
 	type: PreviewType,
     props: Object,
 };